fix: address PR #217 review comments
- sorting correct_index requires int(), so the distinct-of-4 refine
  now actually enforces a permutation of {1,2,3,4}
- record wasted GenerationMeta immediately per attempt instead of a
  spent-minus-kept diff at the end, so a cancellation mid-rewrite no
  longer drops billed calls from the audit trail
- bounded slot loop (MAX_SLOT_ATTEMPTS) instead of hand-unrolled
  first/second attempts
- batch a slot's wasted rows in one recordGenerationMetaMany call
- reviewFeedback consumes the same {name,score,comment} shape
  applicableMetrics already built, no re-lookup by string key

+++ src/types/questions.ts
@@ -88,7 +88,7 @@ export const fillTheGapContentSchema = z.object({
 export const sortingAnswerSchema = z.object({
   id: z.string(),
   text: z.string(),
-  correct_index: z.number().min(1).max(4),
+  correct_index: z.number().int().min(1).max(4),
   explanation: z.string(),
 })
